Update the base image version to 3.8.2. (#205)
As reported there, seemingly this image has the CA file which is
probably old enough so that fetching resources for nodejs
installation fails due to certificate verification.

This primarily update the version to 3.8.2, the latest one.
Some installed package names have change since then; most notably
it's switched to python3. Fortunately the current code seems
working just fine but the command names need to be renamed.